+++ tests/core/tests/test_resources.py
@@ -503,7 +503,7 @@ class ModelResourceTest(TestCase):
         with self.assertWarns(DeprecationWarning) as w:
             self.resource.export(Book.objects.none())
             self.assertEqual(
-                "queryset must be supplied as a named parameter - this will be enforced in a future release",
+                "queryset must be supplied as a named parameter",
                  str(w.warnings[0].message)
             )